wiki_summary/requirements.txt
# Wiki Summary Module Requirements
# Isolated dependencies to avoid conflicts with other projects

# Core dependencies
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# DSPy and LLM
dspy-ai>=2.4.0
litellm>=1.46.0
openai>=1.0.0

# Database
# sqlite3 is built-in to Python

# Data processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0

# Development (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
mypy>=1.5.1
black>=23.9.0